requirements.txt
pytest>=7.1.3
pytest-playwright>=0.3.0
pytest-xdist>=3.2.0
playwright==1.32.1
requests>=2.28.1
allure-pytest>=2.11.1
python-dotenv>=0.21.1
